Fix slim artifact publish: publish struct_type helper classes artifact too
Splitting macro library into compile-time macro and runtime helper classes,
publishing runtime classes as a library for non-fatjar artifact.

Moving compile-time macros to `cognite.spark.compiletime.macros` package.

Verified that compile-time part isn't referenced in artifact
```console
sbt package

for x in `jar tf target/scala-2.12/cdf-spark-datasource_2.12-3.0.0-SNAPSHOT.jar | grep "\.class$" | sed -e 's/\.class$//'`; do
  echo "processing $x";
  javap -c -p -cp target/scala-2.12/cdf-spark-datasource_2.12-3.0.0-SNAPSHOT.jar "$x";
done | grep cognite.spark.compiletime
```

Not adding this a CI check as it is fairly slow and macros part isn't changed too frequently
