+++ src/main/scala/cognite/spark/v1/AssetHierarchyBuilder.scala
@@ -3,7 +3,7 @@ package cognite.spark.v1
 import cats.effect.IO
 import cats.implicits._
 import cognite.spark.v1.PushdownUtilities.stringSeqToCogniteExternalIdSeq
-import cognite.spark.v1.SparkSchemaHelper.{fromRow, structType}
+import cognite.spark.compiletime.macros.SparkSchemaHelper.{fromRow, structType}
 import com.cognite.sdk.scala.common.{CdpApiException, SetValue}
 import com.cognite.sdk.scala.v1.{
   Asset,
@@ -73,6 +73,8 @@ final case class InvalidRootChangeException(assetIds: Seq[String], subtreeId: St
 
 class AssetHierarchyBuilder(config: RelationConfig)(val sqlContext: SQLContext)
     extends CdfRelation(config, "assethierarchy") {
+
+  import cognite.spark.compiletime.macros.StructTypeEncoderMacro._
 
   import CdpConnector.ioRuntime
 
@@ -410,5 +412,7 @@ class AssetHierarchyBuilder(config: RelationConfig)(val sqlContext: SQLContext)
 }
 
 object AssetHierarchyBuilder {
+  import cognite.spark.compiletime.macros.StructTypeEncoderMacro._
+
   val upsertSchema: StructType = structType[AssetsIngestSchema]()
 }
