+++ src/main/scala/cognite/spark/v1/DataModelInstancesHelper.scala
@@ -1,6 +1,6 @@
 package cognite.spark.v1
 
-import cognite.spark.v1.SparkSchemaHelper.fromRow
+import cognite.spark.compiletime.macros.SparkSchemaHelper.fromRow
 
 import java.time.{Instant, LocalDate, LocalDateTime, OffsetDateTime, ZoneId, ZonedDateTime}
 import com.cognite.sdk.scala.v1.DataModelType.NodeType
