+++ src/main/scala/cognite/spark/v1/LabelsRelation.scala
@@ -1,7 +1,7 @@
 package cognite.spark.v1
 
 import cats.effect.IO
-import cognite.spark.v1.SparkSchemaHelper._
+import cognite.spark.compiletime.macros.SparkSchemaHelper._
 import com.cognite.sdk.scala.v1.{GenericClient, Label, LabelCreate, LabelsFilter}
 import org.apache.spark.sql.sources.{Filter, InsertableRelation}
 import org.apache.spark.sql.types._
@@ -13,7 +13,7 @@ class LabelsRelation(config: RelationConfig)(val sqlContext: SQLContext)
     extends SdkV1Relation[Label, String](config, "labels")
     with InsertableRelation
     with WritableRelation {
-
+  import cognite.spark.compiletime.macros.StructTypeEncoderMacro._
   override def schema: StructType = structType[Label]()
 
   override def toRow(a: Label): Row = asRow(a)
@@ -48,6 +48,8 @@ class LabelsRelation(config: RelationConfig)(val sqlContext: SQLContext)
 }
 
 object LabelsRelation {
+  import cognite.spark.compiletime.macros.StructTypeEncoderMacro._
+
   val insertSchema: StructType = structType[LabelInsertSchema]()
   val readSchema: StructType = structType[LabelReadSchema]()
   val deleteSchema: StructType = structType[LabelDeleteSchema]()
