+++ src/main/scala/cognite/spark/v1/SequenceRowsRelation.scala
@@ -2,6 +2,7 @@ package cognite.spark.v1
 
 import cats.effect.IO
 import cats.implicits._
+import cognite.spark.compiletime.macros.SparkSchemaHelper
 import com.cognite.sdk.scala.common.CdpApiException
 import com.cognite.sdk.scala.v1._
 import fs2.Stream
