+++ src/main/scala/cognite/spark/v1/StringDataPointsRelation.scala
@@ -6,7 +6,7 @@ import cognite.spark.v1.PushdownUtilities.{
   pushdownToParameters,
   toPushdownFilterExpression
 }
-import cognite.spark.v1.SparkSchemaHelper.{asRow, fromRow, structType}
+import cognite.spark.compiletime.macros.SparkSchemaHelper.{asRow, fromRow, structType}
 import com.cognite.sdk.scala.common.StringDataPoint
 import fs2.Stream
 import org.apache.spark.rdd.RDD
@@ -100,6 +100,7 @@ class StringDataPointsRelationV1(config: RelationConfig)(override val sqlContext
 }
 
 object StringDataPointsRelation extends UpsertSchema {
+  import cognite.spark.compiletime.macros.StructTypeEncoderMacro._
   // We should use StringDataPointsItem here, but doing that gives the error: "constructor Timestamp encapsulates
   // multiple overloaded alternatives and cannot be treated as a method. Consider invoking
   // `<offending symbol>.asTerm.alternatives` and manually picking the required method" in StructTypeEncoder, probably
