+++ src/main/scala/cognite/spark/v1/ThreeDModelRevisionMappingsRelation.scala
@@ -1,7 +1,7 @@
 package cognite.spark.v1
 
 import cats.effect.IO
-import cognite.spark.v1.SparkSchemaHelper._
+import cognite.spark.compiletime.macros.SparkSchemaHelper._
 import com.cognite.sdk.scala.v1.{GenericClient, ThreeDAssetMapping}
 import fs2.Stream
 import org.apache.spark.sql.sources.Filter
@@ -11,6 +11,8 @@ import org.apache.spark.sql.{Row, SQLContext}
 class ThreeDModelRevisionMappingsRelation(config: RelationConfig, modelId: Long, revisionId: Long)(
     val sqlContext: SQLContext)
     extends SdkV1Relation[ThreeDAssetMapping, String](config, "3dmodelrevisionmappings") {
+  import cognite.spark.compiletime.macros.StructTypeEncoderMacro._
+
   override def schema: StructType = structType[ThreeDAssetMapping]()
 
   override def toRow(t: ThreeDAssetMapping): Row = asRow(t)
