+++ src/main/scala/cognite/spark/v1/ThreeDModelRevisionNodesRelation.scala
@@ -1,7 +1,7 @@
 package cognite.spark.v1
 
 import cats.effect.IO
-import cognite.spark.v1.SparkSchemaHelper._
+import cognite.spark.compiletime.macros.SparkSchemaHelper._
 import com.cognite.sdk.scala.v1.{GenericClient, ThreeDNode}
 import fs2.Stream
 import org.apache.spark.sql.sources.Filter
@@ -11,6 +11,8 @@ import org.apache.spark.sql.{Row, SQLContext}
 class ThreeDModelRevisionNodesRelation(config: RelationConfig, modelId: Long, revisionId: Long)(
     val sqlContext: SQLContext)
     extends SdkV1Relation[ThreeDNode, Long](config, "3dmodelrevisionnodes") {
+  import cognite.spark.compiletime.macros.StructTypeEncoderMacro._
+
   override def schema: StructType = structType[ThreeDNode]()
 
   override def toRow(t: ThreeDNode): Row = asRow(t)
