+++ src/main/scala/cognite/spark/v1/ThreeDModelsRelation.scala
@@ -1,7 +1,7 @@
 package cognite.spark.v1
 
 import cats.effect.IO
-import cognite.spark.v1.SparkSchemaHelper._
+import cognite.spark.compiletime.macros.SparkSchemaHelper._
 import com.cognite.sdk.scala.v1.{GenericClient, ThreeDModel}
 import fs2.Stream
 import org.apache.spark.sql.sources.Filter
@@ -12,7 +12,7 @@ final case class ModelItem(id: Long, name: String, createdTime: Long)
 
 class ThreeDModelsRelation(config: RelationConfig)(val sqlContext: SQLContext)
     extends SdkV1Relation[ThreeDModel, Long](config, "threeDModels.read") {
-
+  import cognite.spark.compiletime.macros.StructTypeEncoderMacro._
   override def schema: StructType = structType[ThreeDModel]()
 
   override def toRow(t: ThreeDModel): Row = asRow(t)
