+++ struct_type/src/main/scala/cognite/spark/v1/StructTypeEncoder.scala
@@ -1,0 +1,7 @@
+package cognite.spark.v1
+
+import org.apache.spark.sql.types.StructType
+
+trait StructTypeEncoder[T] {
+  def structType(): StructType
+}
